Create phone number validator
test your ablity to use regex and verify a string of various telephone numbers. this problem was a little tricky to get started at first but with a bit of regex help I got it, my skills keep improving. but i must not rush the process of learning it seems as if I'm in hurry slow == fast.