cash registar
a function that takes in price and cash and money in the drawer and returns outputs of open, "closed or insufficient funds. it took nearly two days of trying to find adequate solutions to solve this probe but I'm having difficulty with array methods such as reduce I will keep trying to reverse engineer solutions to get a better grasp of the output. but on the other hand, my ability to think through problems computational has made a lot of progress but more work is required always.